Tim2021 Nov

+++ tests/scenarios/PowershellScriptTest/main.cpp
@@ -19,7 +19,7 @@ using namespace std::literals;
 namespace details
 {
     inline std::wstring appmodel_string(LONG(__stdcall *AppModelFunc)(UINT32*, PWSTR))
-    {
+    { 
         // NOTE: `length` includes the null character both as input and output, hence the +1/-1 everywhere
         UINT32 length = MAX_PATH + 1;
         std::wstring result(length - 1, '\0');
@@ -90,50 +90,95 @@ int wmain(int argc, const wchar_t** argv)
 {
     auto result = parse_args(argc, argv);
     std::wstring aumid = details::appmodel_string(&::GetCurrentApplicationUserModelId);
-    test_initialize("Powershell Script Tests", 1);
-    test_begin("Powershell Script Test");
-
     //get rid of the !
     std::wstring testType = aumid.substr(aumid.find('!') + 1);
     std::transform(testType.begin(), testType.end(), testType.begin(), towlower);
 
-    Log(L"<<<<<Powershell Script Test %ls", testType.c_str());
-
     TCHAR localAppDataPath[MAX_PATH];
     SHGetFolderPathW(nullptr, CSIDL_LOCAL_APPDATA, nullptr, SHGFP_TYPE_CURRENT, localAppDataPath);
-    bool doesHelloExist = DoesFileExist(localAppDataPath, L"Hello.txt");
-    bool doesArgumentExist = DoesFileExist(localAppDataPath, L"Argument.txt");
 
-
-    if (testType.compare(L"psonlystart") == 0)
+    if (testType.compare(L"psfshelllaunchtest") == 0)
     {
-        if (!doesHelloExist)
+        // Oh, Just figured out that because we launch a cmd for this test and not this exe,
+        // this exe is not started so this code never gets hit which causes taskrunner to never
+        // see the reporting.
+        test_initialize("PsfShellLaunch Shell Launch Tests", 1);
+        test_begin("PsfShellLaunch Shell Launch Test");
+        Log(L"<<<<<Powershell Shell Launch Test %ls", testType.c_str());
+        bool doesHelloCopyExist = DoesFileExist(localAppDataPath, L"HelloWorldCopy.txt");
+        if (!doesHelloCopyExist)
         {
             result = ERROR_FILE_NOT_FOUND;
         }
+        else
+        {
+           RemoveFile(localAppDataPath, L"HelloWorldCopy.txt");
+        }
+        test_end(result);
+        Log(L"Powershell Shell Launch Test>>>>>");
     }
-    else if (testType.compare(L"psbothstartingfirst") == 0)
+    else if(testType.compare(L"psfshelllaunchverify") == 0)
     {
-        if (!doesHelloExist)
+        test_initialize("PsfShellLaunch Shell Verify Tests", 1);
+        test_begin("PsfShellLaunch Shell Verify Test");
+        Log(L"<<<<<Powershell Shell Launch Test %ls", testType.c_str());
+        bool doesHelloCopyExist = DoesFileExist(localAppDataPath, L"HelloWorldCopy.txt");
+        if (!doesHelloCopyExist)
         {
             result = ERROR_FILE_NOT_FOUND;
         }
-    }
-    else if (testType.compare(L"psscriptwitharg") == 0)
-    {
-        if (!doesArgumentExist)
+        else
         {
-            result = ERROR_FILE_NOT_FOUND;
+            RemoveFile(localAppDataPath, L"HelloWorldCopy.txt");
         }
+        test_end(result);
+        Log(L"Powershell Shell Verify Test>>>>>");
     }
+    else
+    {
+        bool doesHelloExist = DoesFileExist(localAppDataPath, L"Hello.txt");
+        bool doesArgumentExist = DoesFileExist(localAppDataPath, L"Argument.txt");
+
+        test_initialize("Powershell Script Tests", 1);
+        test_begin("Powershell Script Test");
+
+ 
+        Log(L"<<<<<Powershell Script Test %ls", testType.c_str());
 
 
-    RemoveFile(localAppDataPath, L"Hello.txt");
-    RemoveFile(localAppDataPath, L"Argument.txt");
+        if (testType.compare(L"psonlystart") == 0)
+        {
+            if (!doesHelloExist)
+            {
+                result = ERROR_FILE_NOT_FOUND;
+            }
+        }
+        else if (testType.compare(L"psbothstartingfirst") == 0)
+        {
+            if (!doesHelloExist)
+            {
+                result = ERROR_FILE_NOT_FOUND;
+            }
+        }
+        else if (testType.compare(L"psscriptwitharg") == 0)
+        {
+            if (!doesArgumentExist)
+            {
+                result = ERROR_FILE_NOT_FOUND;
+            }
+        }
 
-    test_end(result);
-    Log(L"Powershell Script Test>>>>>");
 
+        if (DoesFileExist(localAppDataPath, L"Hello.txt"))
+            RemoveFile(localAppDataPath, L"Hello.txt");
+        if (DoesFileExist(localAppDataPath, L"Argument.txt"))
+            RemoveFile(localAppDataPath, L"Argument.txt");
+
+        test_end(result);
+        Log(L"Powershell Script Test>>>>>");
+
+
+    }
     test_cleanup();
 
     return result;
